+++ src/components/PhotoCard.vue
